+++ guillotina_stripe/tests/test_subscription.py
@@ -227,8 +227,9 @@ async def test_pay_subscription_us(container_requester):
                 'pmid': pmid
             })
         )
-
         assert resp['status'] == 'active'
+        assert resp["discount"] is None
+        total_amount = resp["items"]["data"][0]["plan"]["amount"]
 
         resp, status_code = await requester(
             "GET",
@@ -237,7 +238,17 @@ async def test_pay_subscription_us(container_requester):
 
         assert resp['subscribed'] is True
 
-        
+        resp, status_code = await requester(
+            "POST",
+            "/db/guillotina/subscription/@subscribe",
+            data=json.dumps({
+                'pmid': pmid,
+                'coupon': "foo-coupon-25"
+            })
+        )
+        assert resp['status'] == 'active'
+        assert isinstance(resp["discount"], dict)
+
 @pytest.mark.asyncio
 async def test_pay_subscription_eu(container_requester):
     async with container_requester as requester: