+++ CHANGELOG.rst
@@ -4,14 +4,18 @@ CHANGELOG
 1.0.0a11 (unreleased)
 ---------------------
 
-- Nothing changed yet.
-
+- Fixing amount in coupons when paying products. Adding tests.
+  [nilbacardit26]
+- When amount to pay using coupons is below 50cts, the total
+  payment would be 50cts, due to slack does not admit payment below 50cts.
+  [nilbacardit26]
 
 1.0.0a10 (2021-11-24)
 ---------------------
 
 - Be able to update customer when calling @register-card, customer_id
   can be passed as a parameter
+  [nilbacardit26]
 
 
 1.0.0a9 (2021-11-24)